requirements.txt
# Home Health AI API Requirements

# Django Core
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-extensions>=3.2.0
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
python-decouple>=3.8

# Database
psycopg2-binary>=2.9.0

# File Processing & OCR
pillow>=10.0.0
pytesseract>=0.3.13
opencv-python>=4.10.0
pdf2image>=1.17.0
python-docx>=1.1.2
openpyxl>=3.1.5
xlsxwriter>=3.2.0
pdfplumber>=0.11.4
pymupdf>=1.24.12

# AI/ML Core
openai>=1.54.4
anthropic>=0.40.0
langchain>=0.3.7
langchain-openai>=0.2.8
langchain-anthropic>=0.2.4

# Machine Learning
transformers>=4.46.2
torch>=2.5.1
scikit-learn>=1.5.2
pandas>=2.2.3
numpy>=2.1.3
sentence-transformers>=3.3.0
huggingface-hub>=0.26.2
datasets>=3.1.0
accelerate>=1.1.1

# Natural Language Processing
nltk>=3.9.1
spacy>=3.8.2
textblob>=0.18.0

# Background Tasks
celery>=5.4.0
redis>=5.2.0
flower>=2.0.1
django-celery-beat>=2.7.0
django-celery-results>=2.5.1

# Utilities
python-multipart>=0.0.6
